requirements.txt
Django>=5.1,<5.2
mysqlclient>=2.1.0
djangorestframework>=3.14.0
